+++ projects/angular-rave/src/lib/angular-rave-token.ts
@@ -1,5 +1,3 @@
 import { InjectionToken } from '@angular/core';
 
-export const PBFPUBKEY_TOKEN = new InjectionToken<string>('Rave.PBFPubKey');
-
-export const ENVIRONMENT_TOKEN = new InjectionToken<string>('Rave.IsDevelopment');
+export const PUBKEY_TOKEN = new InjectionToken<string>('Rave.PubKey');