+++ projects/angular-rave/src/lib/angular-rave.directive.spec.ts
@@ -1,32 +1,77 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { Component, DebugElement } from '@angular/core';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
 
-import { AngularRaveComponent } from './angular-rave.component';
 import { AngularRaveService } from './angular-rave.service';
-import { PBFPUBKEY_TOKEN, ENVIRONMENT_TOKEN } from './angular-rave-token';
+import { PUBKEY_TOKEN } from './angular-rave-token';
+import { AngularRaveDirective } from './angular-rave.directive';
+import { RaveOptions } from './rave-options';
 
-describe('AngularRaveComponent', () => {
-  let component: AngularRaveComponent;
-  let fixture: ComponentFixture<AngularRaveComponent>;
+@Component({
+  template: `
+   <sample-component>
+  `
+})
+class TestComponent { }
 
-  beforeEach(async(() => {
+describe('AngularRaveDirective', () => {
+  let component: any;
+  let fixture: ComponentFixture<any>;
+  let inputEl: DebugElement;
+  const paymentOptions: RaveOptions = {
+    customer: {
+      email:   'someuser@email.com',
+      phonenumber:   '09209209309090'
+    },
+    customizations: {
+      title: 'My title'
+    },
+    currency: 'NGN',
+    amount:  8000,
+    tx_ref:  'A unique transaction reference',
+    paymentOptions: ['account', 'banktransfer', 'card'],
+  };
+
+  beforeEach(waitForAsync(() => {
+    @Component({
+      template: `
+       <button
+        angular-rave
+        [raveOptions]="options"
+        (init)="paymentInit()"
+       >
+       </button>
+      `
+    })
+    class SampleComponent {
+      options = paymentOptions;
+      paymentInit() {}
+    }
+
     TestBed.configureTestingModule({
-      declarations: [ AngularRaveComponent ],
+      declarations: [ TestComponent, AngularRaveDirective, SampleComponent ],
       providers: [
         AngularRaveService,
-        { provide: PBFPUBKEY_TOKEN, useValue: 'PBFPubKey' },
-        { provide: ENVIRONMENT_TOKEN, useValue: true }
+        { provide: PUBKEY_TOKEN, useValue: 'public_key' },
       ]
     })
     .compileComponents();
-  }));
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(AngularRaveComponent);
+    fixture = TestBed.createComponent(SampleComponent);
     component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+    inputEl = fixture.debugElement.query(By.css('button'));
+  }));
 
   it('should create', () => {
     expect(component).toBeTruthy();
   });
+
+  it('should show modal', () => {
+    spyOn(component, 'paymentInit');
+    expect(component).toBeTruthy();
+    setTimeout(() => {
+      expect(component.paymentInit).toHaveBeenCalled();
+    }, 2000);
+  });
+
 });