+++ projects/angular-rave/src/lib/angular-rave.service.spec.ts
@@ -1,15 +1,22 @@
 import { TestBed, inject } from '@angular/core/testing';
 
 import { AngularRaveService } from './angular-rave.service';
-import { PBFPUBKEY_TOKEN, ENVIRONMENT_TOKEN } from './angular-rave-token';
+import { PUBKEY_TOKEN } from './angular-rave-token';
+import { PrivateRaveOptions } from './private-rave-options';
+
+interface MyWindow extends Window {
+  FlutterwaveCheckout: (raveOptions: Partial<PrivateRaveOptions>) => void;
+}
+declare var window: MyWindow;
 
 describe('AngularRaveService', () => {
   beforeEach(() => {
+    window.FlutterwaveCheckout = undefined;
+
     TestBed.configureTestingModule({
       providers: [
         AngularRaveService,
-        { provide: PBFPUBKEY_TOKEN, useValue: 'PBFPubKey' },
-        { provide: ENVIRONMENT_TOKEN, useValue: true }
+        { provide: PUBKEY_TOKEN, useValue: 'PBFPubKey' },
       ]
     });
   });
@@ -18,8 +25,14 @@ describe('AngularRaveService', () => {
     expect(service).toBeTruthy();
   }));
 
-  it('should inject tokens', inject([AngularRaveService], (service: any) => {
-    expect(service.PBFPubKey).toEqual('PBFPubKey');
-    expect(service.isDev).toBeTruthy();
+  it('should inject tokens', inject([AngularRaveService], (service: AngularRaveService) => {
+    expect(service.publicKey).toEqual('PBFPubKey');
+  }));
+
+  it('should load the flutterwave script', inject([AngularRaveService], (service: AngularRaveService) => {
+    expect(typeof window.FlutterwaveCheckout).toEqual('undefined');
+    service.loadScript().then(() => {
+      expect(typeof window.FlutterwaveCheckout).toEqual('function');
+    });
   }));
 });