+++ projects/angular-rave/src/public-api.ts
@@ -5,5 +5,6 @@
 export * from './lib/angular-rave.service';
 export * from './lib/angular-rave.component';
 export * from './lib/angular-rave.directive';
-export { RaveOptions, PaymentInstance } from './lib/rave-options';
+export * from './lib/rave-options';
 export * from './lib/angular-rave.module';
+export { RaveCallback, RavePaymentData } from './lib/private-rave-options';